Signature has invalid URL characters
When attempting to use this module I kept getting HTTP 403 Forbidden
errors from AWS.  Looking into it further and comparing it to the
current AWS CloudFront  documentation at
http://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/private-content-creating-signed-url-canned-policy.html
for creating a Signed URL using a Canned Policy I discovered it
neglected to map the three invalid URL characters "+=/" to "-_~".
Therefore I have updated the code to support this and more closely
conform to the AWS  documentation.